Added exportFormat parameter
Instead of passing .pdf to the ChatView, LLMChatView now has an optional parameter exportFormat that is passed to the ChatView. By default, it is .pdf (so the .init signature need not change), but can now take any ot .pdf, .json, .text, or .none. If exportFormat is .none, no export button will appear in the toolbar.